# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/documents/policy": {
      "get": {
        "operationId": "DocumentsController_policy",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "documents"
        ]
      }
    },
    "/api/v1/documents": {
      "get": {
        "operationId": "DocumentsController_list",
        "parameters": [
          {
            "name": "entityType",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "BUSINESS",
                "DRIVER"
              ]
            }
          },
          {
            "name": "entityId",
            "required": true,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "documents"
        ]
      },
      "post": {
        "operationId": "DocumentsController_upload",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file",
                  "entityType",
                  "entityId",
                  "type"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "entityType": {
                    "type": "string",
                    "enum": [
                      "BUSINESS",
                      "DRIVER"
                    ]
                  },
                  "entityId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "type": {
                    "type": "string"
                  },
                  "expiresAt": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "replacesId": {
                    "type": "string",
                    "format": "uuid"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "documents"
        ]
      }
    },
    "/api/v1/documents/{id}/content": {
      "get": {
        "operationId": "DocumentsController_content",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "documents"
        ]
      }
    },
    "/api/v1/documents/{id}/review": {
      "patch": {
        "operationId": "DocumentsController_review",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDocumentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "documents"
        ]
      }
    },
...
    "/api/v1/routes": {
      "get": {
        "operationId": "RoutesController_list",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "BUILDING",
                "OFFERING",
                "ASSIGNED",
                "ACTIVE",
                "COMPLETED",
                "CANCELLED",
                "FAILED"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "routes"
        ]
      }
    },
    "/api/v1/routes/{id}": {
      "get": {
        "operationId": "RoutesController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Read an authorized route; businesses see only their own deliveries",
        "tags": [
          "routes"
        ]
      }
    },
    "/api/v1/routes/{id}/start": {
      "post": {
        "operationId": "RoutesController_start",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "routes"
        ]
      }
    },
    "/api/v1/routes/{id}/stops/{stopId}/arrive": {
      "post": {
        "operationId": "RoutesController_arrive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "stopId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "routes"
        ]
      }
    },
    "/api/v1/routes/{id}/stops/{stopId}/complete": {
      "post": {
        "operationId": "RoutesController_complete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "stopId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Complete pickup; dropoff completion requires recipient proof",
        "tags": [
          "routes"
        ]
      }
    },
    "/api/v1/routes/{id}/finish": {
      "post": {
        "operationId": "RoutesController_finish",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "routes"
        ]
      }
    },
...
    "/api/v1/whatsapp/me": {
      "get": {
        "operationId": "WhatsAppAccountController_status",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "whatsapp"
        ]
      }
    },
    "/api/v1/whatsapp/me/consent": {
      "post": {
        "operationId": "WhatsAppAccountController_consent",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WhatsAppConsentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "whatsapp"
        ]
      }
    },
...
  },
  "info": {
    "title": "ORBITA",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ReviewDocumentDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "APPROVED",
              "REJECTED"
            ]
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "notes"
        ]
      },
...
      "WhatsAppConsentDto": {
        "type": "object",
        "properties": {
          "granted": {
            "type": "boolean"
          }
        },
        "required": [
          "granted"
        ]
      },
...
    }
  }
}